+++ point/src/main/java/com/sparta/point/domain/model/PointHistory.java
@@ -1,4 +1,4 @@
-package com.sparta.point.domain;
+package com.sparta.point.domain.model;
 
 import jakarta.persistence.*;
 import lombok.AccessLevel;
